2348. Number of Zero-Filled Subarrays.
